+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ Release Notes
     * [SER-1544]-Office dropdown in Create/Edit Client interface should be searchable
     * [SER-1543]-Office dropdown in Import Client interface should be searchable
     * [SER-1541]-Office dropdownlist in the transfer client interface should be searchable
+    * [SER-1569]-Check how transfer client endpoint works
 
 ## Version 1.0.0 - for use with Fineract Web App